% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiname.R
\name{get_multinames}
\alias{get_multinames}
\title{Get Assigned Multinames}
\usage{
get_multinames(.data, from = NULL)
}
\arguments{
\item{.data}{An initialized data.frame, tibble, or data.table}

\item{from}{Specifies which multinames to collect. If `NULL`, then a data.frame with all multinames is returned.}
}
\description{
Get Assigned Multinames
}
\examples{
df <- initialize_multinames(data.frame(a=1:10, b=11:20))
add_multinames(df, c("a" = "A", "b" = "B"), "CAPS")
get_multinames(df)           # Returns all multinames in a data.frame
get_multinames(df, "CAPS")   # Returns just multinames in "CAPS"
}
