% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiname.R
\name{multiname}
\alias{multiname}
\title{Multiname Backend}
\description{
Multiname Backend

Multiname Backend
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{multiname$new()}}
\item \href{#method-get_multinames}{\code{multiname$get_multinames()}}
\item \href{#method-set_multinames}{\code{multiname$set_multinames()}}
\item \href{#method-add_multinames}{\code{multiname$add_multinames()}}
\item \href{#method-replace_multinames}{\code{multiname$replace_multinames()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiname$new(.data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_multinames"></a>}}
\if{latex}{\out{\hypertarget{method-get_multinames}{}}}
\subsection{Method \code{get_multinames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiname$get_multinames(from = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_multinames"></a>}}
\if{latex}{\out{\hypertarget{method-set_multinames}{}}}
\subsection{Method \code{set_multinames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiname$set_multinames(.data, to = colnames(private$.multinames))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_multinames"></a>}}
\if{latex}{\out{\hypertarget{method-add_multinames}{}}}
\subsection{Method \code{add_multinames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiname$add_multinames(named_vector, to, from = NULL, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_multinames"></a>}}
\if{latex}{\out{\hypertarget{method-replace_multinames}{}}}
\subsection{Method \code{replace_multinames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiname$replace_multinames(
  named_vector,
  to = colnames(private$.multinames),
  from = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
}
