% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiname.R
\name{set_multinames}
\alias{set_multinames}
\title{Set Assigned Multinames}
\usage{
set_multinames(.data, to)
}
\arguments{
\item{.data}{An initialized data.frame, tibble, or data.table}

\item{to}{Specifies which multinames to collect.}
}
\description{
Set Assigned Multinames
}
\examples{
df <- initialize_multinames(data.frame(a=1:10, b=11:20))
add_multinames(df, c("a" = "A", "b" = "B"), "CAPS")
set_multinames(df, "CAPS") # Returns df with names specified in "CAPS"
}
